Create settings sub-module
