+++ dev/App.tsx
@@ -3,8 +3,8 @@ import { FetchHttpClient } from '@shlinkio/shlink-js-sdk/browser';
 import type { FC } from 'react';
 import { useEffect, useMemo, useState } from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
-import type { Settings } from '../src';
 import { ShlinkWebComponent } from '../src';
+import type { Settings } from '../src/settings';
 import type { SemVer } from '../src/utils/helpers/version';
 import { ServerInfoForm } from './server-info/ServerInfoForm';
 import type { ServerInfo } from './server-info/useServerInfo';
