+++ package.json
@@ -5,7 +5,7 @@
   "repository": "https://github.com/shlinkio/shlink-web-component",
   "license": "MIT",
   "type": "module",
-  "main": "./dist/index.umd.cjs",
+  "main": "./dist/index.cjs",
   "module": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "exports": {
@@ -18,6 +18,11 @@
       "import": "./dist/api-contract.js",
       "require": "./dist/api-contract.cjs",
       "types": "./dist/api-contract.d.ts"
+    },
+    "./settings": {
+      "import": "./dist/settings.js",
+      "require": "./dist/settings.cjs",
+      "types": "./dist/settings.d.ts"
     },
     "./package.json": "./package.json"
   },
