+++ src/ShlinkWebComponent.tsx
@@ -5,12 +5,12 @@ import { Fragment, useEffect, useMemo, useRef, useState } from 'react';
 import { Provider as ReduxStoreProvider } from 'react-redux';
 import { BrowserRouter, useInRouterContext } from 'react-router-dom';
 import type { ShlinkApiClient } from './api-contract';
+import type { Settings } from './settings';
+import { SettingsProvider } from './settings';
 import { FeaturesProvider, useFeatures } from './utils/features';
 import type { SemVerOrLatest } from './utils/helpers/version';
 import { RoutesPrefixProvider } from './utils/routesPrefix';
 import type { TagColorsStorage } from './utils/services/TagColorsStorage';
-import type { Settings } from './utils/settings';
-import { SettingsProvider } from './utils/settings';
 
 type ShlinkWebComponentProps = {
   serverVersion: SemVerOrLatest; // FIXME Consider making this optional and trying to resolve it if not set
