+++ src/index.ts
@@ -10,13 +10,4 @@ export const ShlinkWebComponent = createShlinkWebComponent(bottle);
 
 export type ShlinkWebComponentType = typeof ShlinkWebComponent;
 
-export type {
-  RealTimeUpdatesSettings,
-  ShortUrlCreationSettings,
-  ShortUrlsListSettings,
-  VisitsSettings,
-  TagsSettings,
-  Settings,
-} from './utils/settings';
-
 export type { TagColorsStorage } from './utils/services/TagColorsStorage';
