+++ src/short-urls/CreateShortUrl.tsx
@@ -3,8 +3,8 @@ import { useMemo } from 'react';
 import type { ShlinkCreateShortUrlData } from '../api-contract';
 import type { FCWithDeps } from '../container/utils';
 import { componentFactory, useDependencies } from '../container/utils';
-import type { ShortUrlCreationSettings } from '../utils/settings';
-import { useSetting } from '../utils/settings';
+import type { ShortUrlCreationSettings } from '../settings';
+import { useSetting } from '../settings';
 import type { CreateShortUrlResultProps } from './helpers/CreateShortUrlResult';
 import type { ShortUrlCreation } from './reducers/shortUrlCreation';
 import type { ShortUrlFormProps } from './ShortUrlForm';
