+++ src/short-urls/helpers/ShortUrlsRow.tsx
@@ -4,10 +4,10 @@ import { ExternalLink } from 'react-external-link';
 import type { ShlinkShortUrl } from '../../api-contract';
 import type { FCWithDeps } from '../../container/utils';
 import { componentFactory, useDependencies } from '../../container/utils';
+import { useSetting } from '../../settings';
 import { CopyToClipboardIcon } from '../../utils/components/CopyToClipboardIcon';
 import { Time } from '../../utils/dates/Time';
 import type { ColorGenerator } from '../../utils/services/ColorGenerator';
-import { useSetting } from '../../utils/settings';
 import { useShortUrlsQuery } from './hooks';
 import type { ShortUrlsRowMenuType } from './ShortUrlsRowMenu';
 import { ShortUrlStatus } from './ShortUrlStatus';
