--- src/utils/settings.ts
@@ -1,77 +1,0 @@
-import { createContext, useContext } from 'react';
-import type { ShortUrlsOrder } from '../short-urls/data';
-import type { TagsOrder } from '../tags/data/TagsListChildrenProps';
-import type { DateInterval } from './dates/helpers/dateIntervals';
-
-export const DEFAULT_SHORT_URLS_ORDERING: ShortUrlsOrder = {
-  field: 'dateCreated',
-  dir: 'DESC',
-};
-
-/**
- * Important! When adding new props in the main Settings interface or any of the nested props, they have to be set as
- * optional, as old instances of the app will load partial objects from local storage until it is saved again.
- */
-
-export type RealTimeUpdatesSettings = {
-  enabled: boolean;
-  interval?: number;
-};
-
-export type TagFilteringMode = 'startsWith' | 'includes';
-
-export type ShortUrlCreationSettings = {
-  tagFilteringMode?: TagFilteringMode;
-  forwardQuery?: boolean;
-
-  /** @deprecated Shlink 4.0.0 no longer validates URLs */
-  validateUrls: boolean;
-};
-
-export type VisitsSettings = {
-  defaultInterval: DateInterval;
-  excludeBots?: boolean;
-  loadPrevInterval?: boolean;
-};
-
-export type TagsSettings = {
-  defaultOrdering?: TagsOrder;
-};
-
-export type ShortUrlsListSettings = {
-  defaultOrdering?: ShortUrlsOrder;
-};
-
-export type Settings = {
-  realTimeUpdates?: RealTimeUpdatesSettings;
-  shortUrlCreation?: ShortUrlCreationSettings;
-  shortUrlsList?: ShortUrlsListSettings;
-  visits?: VisitsSettings;
-  tags?: TagsSettings;
-};
-
-const defaultSettings: Settings = {
-  realTimeUpdates: {
-    enabled: true,
-  },
-  shortUrlCreation: {
-    validateUrls: false,
-  },
-  visits: {
-    defaultInterval: 'last30Days',
-  },
-  shortUrlsList: {
-    defaultOrdering: DEFAULT_SHORT_URLS_ORDERING,
-  },
-};
-
-const SettingsContext = createContext<Settings | undefined>(defaultSettings);
-
-export const SettingsProvider = SettingsContext.Provider;
-
-export const useSettings = (): Settings => useContext(SettingsContext) ?? defaultSettings;
-
-export const useSetting = <T extends keyof Settings>(settingName: T): Settings[T] => {
-  const settings = useSettings();
-  return settings[settingName];
-};
