+++ test/short-urls/helpers/ShortUrlsRow.test.tsx
@@ -3,11 +3,11 @@ import { screen } from '@testing-library/react';
 import { fromPartial } from '@total-typescript/shoehorn';
 import { addDays, formatISO, subDays } from 'date-fns';
 import { MemoryRouter } from 'react-router-dom';
-import type { Settings } from '../../../src';
 import type { ShlinkShortUrl, ShlinkShortUrlMeta } from '../../../src/api-contract';
+import type { Settings } from '../../../src/settings';
+import { SettingsProvider } from '../../../src/settings';
 import { ShortUrlsRowFactory } from '../../../src/short-urls/helpers/ShortUrlsRow';
 import { now, parseDate } from '../../../src/utils/dates/helpers/date';
-import { SettingsProvider } from '../../../src/utils/settings';
 import { checkAccessibility } from '../../__helpers__/accessibility';
 import { renderWithEvents } from '../../__helpers__/setUpTest';
 import { colorGeneratorMock } from '../../utils/services/__mocks__/ColorGenerator.mock';
