+++ test/visits/NonOrphanVisits.test.tsx
@@ -3,7 +3,7 @@ import { fromPartial } from '@total-typescript/shoehorn';
 import { formatISO } from 'date-fns';
 import { MemoryRouter } from 'react-router-dom';
 import type { MercureBoundProps } from '../../src/mercure/helpers/boundToMercureHub';
-import { SettingsProvider } from '../../src/utils/settings';
+import { SettingsProvider } from '../../src/settings';
 import { NonOrphanVisitsFactory } from '../../src/visits/NonOrphanVisits';
 import type { VisitsInfo } from '../../src/visits/reducers/types';
 import { checkAccessibility } from '../__helpers__/accessibility';
