+++ test/visits/OrphanVisits.test.tsx
@@ -3,8 +3,8 @@ import { fromPartial } from '@total-typescript/shoehorn';
 import { formatISO } from 'date-fns';
 import { MemoryRouter } from 'react-router-dom';
 import type { MercureBoundProps } from '../../src/mercure/helpers/boundToMercureHub';
+import { SettingsProvider } from '../../src/settings';
 import { FeaturesProvider } from '../../src/utils/features';
-import { SettingsProvider } from '../../src/utils/settings';
 import { OrphanVisitsFactory } from '../../src/visits/OrphanVisits';
 import type { VisitsInfo } from '../../src/visits/reducers/types';
 import { checkAccessibility } from '../__helpers__/accessibility';
