+++ test/visits/ShortUrlVisits.test.tsx
@@ -5,8 +5,8 @@ import { formatISO } from 'date-fns';
 import { MemoryRouter } from 'react-router-dom';
 import { now } from 'tinybench';
 import type { MercureBoundProps } from '../../src/mercure/helpers/boundToMercureHub';
+import { SettingsProvider } from '../../src/settings';
 import { FeaturesProvider } from '../../src/utils/features';
-import { SettingsProvider } from '../../src/utils/settings';
 import type { ShortUrlVisits as ShortUrlVisitsState } from '../../src/visits/reducers/shortUrlVisits';
 import { ShortUrlVisitsFactory } from '../../src/visits/ShortUrlVisits';
 import { checkAccessibility } from '../__helpers__/accessibility';
