+++ test/visits/TagVisits.test.tsx
@@ -2,7 +2,7 @@ import { screen } from '@testing-library/react';
 import { fromPartial } from '@total-typescript/shoehorn';
 import { formatISO } from 'date-fns';
 import type { MercureBoundProps } from '../../src/mercure/helpers/boundToMercureHub';
-import { SettingsProvider } from '../../src/utils/settings';
+import { SettingsProvider } from '../../src/settings';
 import type { TagVisits as TagVisitsStats } from '../../src/visits/reducers/tagVisits';
 import type { TagVisitsProps } from '../../src/visits/TagVisits';
 import { TagVisitsFactory } from '../../src/visits/TagVisits';
