+++ test/visits/VisitsStats.test.tsx
@@ -2,10 +2,10 @@ import { screen, waitFor } from '@testing-library/react';
 import { fromPartial } from '@total-typescript/shoehorn';
 import { createMemoryHistory } from 'history';
 import { Router } from 'react-router-dom';
-import type { Settings } from '../../src';
 import type { ShlinkVisit } from '../../src/api-contract';
+import type { Settings } from '../../src/settings';
+import { SettingsProvider } from '../../src/settings';
 import { rangeOf } from '../../src/utils/helpers';
-import { SettingsProvider } from '../../src/utils/settings';
 import type { VisitsInfo } from '../../src/visits/reducers/types';
 import { VisitsStats } from '../../src/visits/VisitsStats';
 import { checkAccessibility } from '../__helpers__/accessibility';
